5354 Rename sleeper.parquet.row package
