+++ java/athena/src/main/java/sleeper/athena/record/SimpleRecordHandler.java
@@ -31,8 +31,8 @@ import sleeper.core.iterator.CloseableIterator;
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.row.Row;
 import sleeper.core.schema.Schema;
-import sleeper.parquet.record.ParquetReaderIterator;
-import sleeper.parquet.record.ParquetRowReader;
+import sleeper.parquet.row.ParquetReaderIterator;
+import sleeper.parquet.row.ParquetRowReader;
 
 import java.util.List;
 import java.util.Set;
