+++ java/bulk-import/bulk-import-runner/src/main/java/sleeper/bulkimport/runner/dataframe/FileWritingIterator.java
@@ -31,7 +31,7 @@ import sleeper.core.schema.type.ListType;
 import sleeper.core.schema.type.MapType;
 import sleeper.core.table.TableFilePaths;
 import sleeper.core.util.LoggedDuration;
-import sleeper.parquet.record.ParquetRecordWriterFactory;
+import sleeper.parquet.row.ParquetRecordWriterFactory;
 import sleeper.sketches.Sketches;
 import sleeper.sketches.store.SketchesStore;
