+++ java/bulk-import/bulk-import-runner/src/main/java/sleeper/bulkimport/runner/rdd/SingleFileWritingIterator.java
@@ -34,7 +34,7 @@ import sleeper.core.schema.type.ListType;
 import sleeper.core.schema.type.MapType;
 import sleeper.core.table.TableFilePaths;
 import sleeper.core.util.LoggedDuration;
-import sleeper.parquet.record.ParquetRecordWriterFactory;
+import sleeper.parquet.row.ParquetRecordWriterFactory;
 import sleeper.sketches.Sketches;
 import sleeper.sketches.store.SketchesStore;
