+++ java/clients/src/main/java/sleeper/clients/table/partition/EstimateSplitPointsClient.java
@@ -23,8 +23,8 @@ import sleeper.core.iterator.LimitingIterator;
 import sleeper.core.row.Row;
 import sleeper.core.schema.Schema;
 import sleeper.core.schema.SchemaSerDe;
-import sleeper.parquet.record.ParquetReaderIterator;
-import sleeper.parquet.record.ParquetRowReader;
+import sleeper.parquet.row.ParquetReaderIterator;
+import sleeper.parquet.row.ParquetRowReader;
 import sleeper.parquet.utils.HadoopConfigurationProvider;
 
 import java.io.BufferedWriter;
