+++ java/clients/src/test/java/sleeper/clients/table/partition/EstimateSplitPointsClientIT.java
@@ -25,7 +25,7 @@ import sleeper.core.schema.Schema;
 import sleeper.core.schema.type.IntType;
 import sleeper.core.schema.type.LongType;
 import sleeper.localstack.test.LocalStackTestBase;
-import sleeper.parquet.record.ParquetRecordWriterFactory;
+import sleeper.parquet.row.ParquetRecordWriterFactory;
 
 import java.io.IOException;
 import java.util.ArrayList;
