+++ java/compaction/compaction-datafusion/src/test/java/sleeper/compaction/datafusion/DataFusionCompactionRunnerLocalStackIT.java
@@ -39,9 +39,9 @@ import sleeper.core.table.TableFilePaths;
 import sleeper.core.tracker.job.run.RowsProcessed;
 import sleeper.localstack.test.LocalStackTestBase;
 import sleeper.localstack.test.SleeperLocalStackContainer;
-import sleeper.parquet.record.ParquetReaderIterator;
-import sleeper.parquet.record.ParquetRecordWriterFactory;
-import sleeper.parquet.record.RowReadSupport;
+import sleeper.parquet.row.ParquetReaderIterator;
+import sleeper.parquet.row.ParquetRecordWriterFactory;
+import sleeper.parquet.row.RowReadSupport;
 import sleeper.sketches.Sketches;
 import sleeper.sketches.store.S3SketchesStore;
 import sleeper.sketches.store.SketchesStore;
