+++ java/ingest/ingest-runner/src/main/java/sleeper/ingest/runner/impl/ParquetConfiguration.java
@@ -21,7 +21,7 @@ import org.apache.parquet.hadoop.ParquetWriter;
 
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.row.Row;
-import sleeper.parquet.record.ParquetRecordWriterFactory;
+import sleeper.parquet.row.ParquetRecordWriterFactory;
 
 import java.io.IOException;
 import java.util.Objects;
