+++ java/ingest/ingest-runner/src/main/java/sleeper/ingest/runner/impl/rowbatch/arraylist/ArrayListRowBatch.java
@@ -31,8 +31,8 @@ import sleeper.core.schema.Schema;
 import sleeper.core.util.LoggedDuration;
 import sleeper.ingest.runner.impl.ParquetConfiguration;
 import sleeper.ingest.runner.impl.rowbatch.RowBatch;
-import sleeper.parquet.record.ParquetReaderIterator;
-import sleeper.parquet.record.ParquetRowReader;
+import sleeper.parquet.row.ParquetReaderIterator;
+import sleeper.parquet.row.ParquetRowReader;
 
 import java.io.IOException;
 import java.text.DecimalFormat;
