+++ java/ingest/ingest-runner/src/test/java/sleeper/ingest/runner/testutils/ResultVerifier.java
@@ -23,8 +23,8 @@ import sleeper.core.iterator.CloseableIterator;
 import sleeper.core.row.Row;
 import sleeper.core.schema.Schema;
 import sleeper.core.statestore.FileReference;
-import sleeper.parquet.record.ParquetReaderIterator;
-import sleeper.parquet.record.ParquetRowReader;
+import sleeper.parquet.row.ParquetReaderIterator;
+import sleeper.parquet.row.ParquetRowReader;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
