+++ java/parquet/src/main/java/sleeper/parquet/row/RowConverter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.parquet.record;
+package sleeper.parquet.row;
 
 import org.apache.parquet.io.api.Binary;
 import org.apache.parquet.io.api.Converter;
