+++ java/system-test/system-test-drivers/src/main/java/sleeper/systemtest/drivers/ingest/HadoopLocalFilesDriver.java
@@ -20,7 +20,7 @@ import org.apache.parquet.hadoop.ParquetWriter;
 
 import sleeper.core.properties.table.TableProperties;
 import sleeper.core.row.Row;
-import sleeper.parquet.record.ParquetRecordWriterFactory;
+import sleeper.parquet.row.ParquetRecordWriterFactory;
 import sleeper.systemtest.dsl.sourcedata.IngestLocalFilesDriver;
 
 import java.io.IOException;
